Add vim tests and better ctrlp matching
This adds vim-test for better testing of the files we're in.
We've also added a better matcher for ctrlp to increase the
speed. I'm hoping this will get my close to the commandT like speed that
I'm used to.

Oh and I'm writing a long commit message so I can see how it renders